Hoist shared hook-test helpers into testFixtures
testHandle, settledData, and makeWrapper were copy-pasted across the
three messages hook test files.
