+++ apps/inspect/src/log_data/messageRowsQuery.test.ts
@@ -1,32 +1,17 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
-import { createElement, ReactNode } from "react";
 import { describe, expect, it } from "vitest";
-
-import { ChatMessage } from "@tsmono/inspect-common/types";
-
-import { SampleHandle } from "../app/types";
 
 import { useMessageRows } from "./messageRowsQuery";
 import { type EvalSampleData } from "./sampleData";
 import {
   failingSequenceReader,
+  testHandle as handle,
   testMessages as makeMessages,
+  makeWrapper,
   sequenceReaderOver,
+  settledData,
   testChunkedSample,
-  testEvalSample,
 } from "./testFixtures";
-
-const handle: SampleHandle = { logFile: "log.eval", id: "s1", epoch: 1 };
-
-const settledData = (messages: ChatMessage[]): EvalSampleData => ({
-  sample: testEvalSample(messages),
-  status: "ok",
-  error: undefined,
-  running: [],
-  eventsCleared: false,
-  backfilling: false,
-});
 
 const streamingData: EvalSampleData = {
   sample: undefined,
@@ -36,11 +21,6 @@ const streamingData: EvalSampleData = {
   eventsCleared: false,
   backfilling: false,
 };
-
-const makeWrapper = (client: QueryClient = new QueryClient()) =>
-  function Wrapper({ children }: { children: ReactNode }) {
-    return createElement(QueryClientProvider, { client }, children);
-  };
 
 const renderRows = (data: EvalSampleData, activated = true) =>
   renderHook(
