+++ apps/inspect/src/log_data/messagesExport.test.ts
@@ -1,36 +1,16 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { renderHook } from "@testing-library/react";
-import { createElement, ReactNode } from "react";
 import { describe, expect, it, vi } from "vitest";
-
-import { ChatMessage } from "@tsmono/inspect-common/types";
-
-import { SampleHandle } from "../app/types";
 
 import { useMessagesExport } from "./messagesExport";
 import { type EvalSampleData } from "./sampleData";
 import {
+  testHandle as handle,
   testMessages as makeMessages,
+  makeWrapper,
   sequenceReaderOver,
+  settledData,
   testChunkedSample,
-  testEvalSample,
 } from "./testFixtures";
-
-const handle: SampleHandle = { logFile: "log.eval", id: "s1", epoch: 1 };
-
-const settledData = (messages: ChatMessage[]): EvalSampleData => ({
-  sample: testEvalSample(messages),
-  status: "ok",
-  error: undefined,
-  running: [],
-  eventsCleared: false,
-  backfilling: false,
-});
-
-const makeWrapper = (client: QueryClient = new QueryClient()) =>
-  function Wrapper({ children }: { children: ReactNode }) {
-    return createElement(QueryClientProvider, { client }, children);
-  };
 
 describe("useMessagesExport", () => {
   it("exports a settled monolith conversation", async () => {
